backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "boto3>=1.38.29",
    "celery>=5.4.0",
    "certbot>=3.3.0",
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-silk>=5.3.2",
    "django-storages>=1.14.6",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.14.1",
    "drf-standardized-response>=1.1.3",
    "flower>=2.0.1",
    "geoip2>=5.0.1",
    "model-bakery>=1.20.4",
    "moviepy>=2.2.1",
    "psycopg[binary]>=3.2.6",
    "pyopenssl>=25.0.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "yandex>=0.0.2",
    "yandexcloud>=0.350.0",
]
